Move AliasIndexer into its own file
